Fix PAC file generation to include user-configured excluded hosts; break down Set into smaller sub-routines
